BigInteger is used to hold large numbers

+++ RMItest/src/IServer.java
@@ -1,7 +1,8 @@
+import java.math.BigInteger;
 import java.rmi.Remote;
 import java.rmi.RemoteException;
 
 public interface IServer extends Remote {
 	public void m() throws RemoteException;
-	public boolean isPrime(int number) throws RemoteException;
+	public boolean isPrime(BigInteger number) throws RemoteException;
 }
